chore(conftest): improve pytest hooks for logs
This commit improves the pytest hooks to create a unique datetime for
the test execution session because before it was recreating for each
test case and if session took more than 1 minute, a new log directory
were created. Also, improve hooks to only creates logfile if skip
mark is not present and to remove any logfile related to a test case
that doesn't have any logging call, which means, file is empty.